+++ app/controllers/AddressController.scala
@@ -136,70 +136,6 @@ class AddressController @Inject()(
       }
     }
 
-  def showUpdateInternationalAddressForm(typ: AddrType): Action[AnyContent] =
-    authenticate.async { implicit request =>
-      gettingCachedJourneyData[Result](typ) { journeyData =>
-        addressJourneyEnforcer { _ => personDetails =>
-          typ match {
-            case PostalAddrType =>
-              auditConnector.sendEvent(
-                buildAddressChangeEvent("postalAddressChangeLinkClicked", personDetails, isInternationalAddress = true))
-              enforceDisplayAddressPageVisited(journeyData.addressPageVisitedDto) {
-                Future.successful(
-                  Ok(
-                    views.html.personaldetails.updateInternationalAddress(
-                      journeyData.submittedAddressDto.fold(AddressDto.internationalForm)(
-                        AddressDto.internationalForm.fill),
-                      typ,
-                      countryHelper.countries
-                    )
-                  )
-                )
-              }
-
-            case _ =>
-              auditConnector.sendEvent(
-                buildAddressChangeEvent("mainAddressChangeLinkClicked", personDetails, isInternationalAddress = true))
-              enforceResidencyChoiceSubmitted(journeyData) { _ =>
-                Future.successful(
-                  Ok(
-                    views.html.personaldetails
-                      .updateInternationalAddress(AddressDto.internationalForm, typ, countryHelper.countries)
-                  )
-                )
-              }
-          }
-        }
-      }
-    }
-
-  def processUpdateInternationalAddressForm(typ: AddrType): Action[AnyContent] =
-    authenticate.async { implicit request =>
-      gettingCachedJourneyData[Result](typ) { _ =>
-        addressJourneyEnforcer { _ => _ =>
-          {
-            AddressDto.internationalForm.bindFromRequest.fold(
-              formWithErrors => {
-                Future.successful(BadRequest(
-                  views.html.personaldetails.updateInternationalAddress(formWithErrors, typ, countryHelper.countries)))
-              },
-              addressDto => {
-                addToCache(SubmittedAddressDtoId(typ), addressDto) flatMap { _ =>
-                  typ match {
-                    case PostalAddrType =>
-                      addToCache(SubmittedStartDateId(typ), DateDto(LocalDate.now()))
-                      Future.successful(Redirect(routes.AddressController.reviewChanges(typ)))
-                    case _ =>
-                      Future.successful(Redirect(routes.AddressController.enterStartDate(typ)))
-                  }
-                }
-              }
-            )
-          }
-        }
-      }
-    }
-
   def nonPostalJourneyEnforcer(typ: AddrType)(block: => Future[Result]): Future[Result] =
     typ match {
       case _: ResidentialAddrType => block
