+++ conf/prod.routes
@@ -49,8 +49,8 @@ GET         /personal-account/your-address/:typ/select-address                  
 POST        /personal-account/your-address/:typ/select-address                         controllers.address.AddressSelectorController.onSubmit(typ: AddrType)
 GET         /personal-account/your-address/:typ/edit-address                           controllers.address.UpdateAddressController.onPageLoad(typ: AddrType)
 POST        /personal-account/your-address/:typ/edit-address                           controllers.address.UpdateAddressController.onSubmit(typ: AddrType)
-GET         /personal-account/your-address/:typ/enter-international-address            controllers.AddressController.showUpdateInternationalAddressForm(typ: AddrType)
-POST        /personal-account/your-address/:typ/enter-international-address            controllers.AddressController.processUpdateInternationalAddressForm(typ: AddrType)
+GET         /personal-account/your-address/:typ/enter-international-address            controllers.address.UpdateInternationalAddressController.onPageLoad(typ: AddrType)
+POST        /personal-account/your-address/:typ/enter-international-address            controllers.address.UpdateInternationalAddressController.onSubmit(typ: AddrType)
 GET         /personal-account/your-address/:typ/enter-start-date                       controllers.AddressController.enterStartDate(typ: AddrType)
 POST        /personal-account/your-address/:typ/enter-start-date                       controllers.AddressController.processEnterStartDate(typ: AddrType)
 GET         /personal-account/your-address/:typ/changes                                controllers.AddressController.reviewChanges(typ: AddrType)
